lifecycle: migrate internal healthcheck to use go
